+++ CHANGELOG.md
@@ -18,7 +18,7 @@ These are the section headers that we use:
 
 ### Added
 
-- Add support for creating and listing questions with `SpansQuestionSettings`. ([#24](https://github.com/argilla-io/argilla-server/pull/24))
+- Add support for creating and listing questions with `SpanQuestionSettings`. ([#24](https://github.com/argilla-io/argilla-server/pull/24))
 
 ## [1.24.0](https://github.com/argilla-io/argilla-server/releases/tag/v1.24.0)